requirements.txt
pip==23.1.2
setuptools==67.8.0
wheel==0.40.0
selenium==4.9.1
pytest==7.3.1
pre-commit==3.3.2
pytest-xdist==3.3.1
websocket==0.2.1
requests==2.31.0
cerberus==1.3.4
allure-pytest==2.13.2